+++ app/routes/docs.circulars.markdown.mdx
@@ -5,7 +5,7 @@ handle:
 
 import { Table } from '@trussworks/react-uswds'
 
-import { SyntaxReference } from '~/routes/circulars.edit.$circularId/CircularEditForm.tsx'
+import { SyntaxReference } from '~/routes/circulars.new.$circularId/CircularEditForm.tsx'
 
 # Astro Flavored Markdown
